// DXC_Frontend/src/api/models/apiResultOfListUploadFileResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { UploadFileResponseDto } from './uploadFileResponseDto'

/**
 * Represents a standard API response for single-item operations (Get, Create, Update, Delete).
 */
export interface ApiResultOfListUploadFileResponseDto {
  /** Indicates whether the request was successful. */
  success?: boolean
  /**
   * The data payload. Can be null for operations like Delete.
   * @nullable
   */
  data?: UploadFileResponseDto[] | null
  /**
   * An optional message, providing more details about the result.
   * @nullable
   */
  message?: string | null
}